CL bugfixing
git-svn-id: http://encog-cs.googlecode.com/svn/trunk/encog-core@315 0d3f3ea6-ac51-0410-a1de-edb5cd2a854a

+++ encog-core-cs/Neural/Networks/Flat/IFlatGradientWorker.cs
@@ -9,6 +9,6 @@ namespace Encog.Neural.Networks.Flat
     public interface IFlatGradientWorker: IEncogTask
     {
         double[] Weights { get;  }
-        int ElapsedTime { get; }
+        long ElapsedTime { get; }
     }
 }
